No longer attempts database retrieval if a connection has not been established.
